pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cnxtvit-age-estimation"
version = "0.1.0"
description = "ConvNeXt + Transformer hybrid for age estimation (research/educational release)"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name = "Rony"}]

dependencies = [
  "torch>=2.0",
  "torchvision>=0.15",
  "pytorch-lightning>=2.0",
  "torchmetrics>=1.0",
  "timm>=0.9",
  "einops>=0.7",
  "opencv-python>=4.8",
  "pandas>=2.0",
  "numpy>=1.23",
  "Pillow>=9.5",
  "yacs>=0.1.8",
  "pyyaml>=6.0",
  "tqdm>=4.66",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
